# web/docs/swagger.yaml
basePath: /
definitions:
  errorinterface.ErrorResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
...
  models.Action:
    properties:
      action:
        type: string
      actions:
        items:
          $ref: '#/definitions/models.Action'
        type: array
      description:
        type: string
    type: object
  models.AuditLog:
    properties:
      action:
        description: Action taken (e.g., "Created Task", "Updated Milestone")
        type: string
      details:
        description: Detailed description of the action
        type: string
      timestamp:
        description: When the action was performed
        type: string
      user_id:
        description: The user who performed the action
        type: string
    type: object
  models.Permission:
    properties:
      actions:
        items:
          $ref: '#/definitions/models.Action'
        type: array
    type: object
...
  roles_permissions.RoleRequest:
    properties:
      audit_logs:
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      created_timestamp_utc:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: string
      owner_id:
        type: string
      permissions:
        additionalProperties:
          $ref: '#/definitions/models.Permission'
        type: object
      project_id:
        type: string
      role:
        type: string
      updated_timestamp_utc:
        type: string
    type: object
  roles_permissions.RoleResponse:
    properties:
      audit_logs:
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      created_timestamp_utc:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: string
      owner_id:
        type: string
      permissions:
        additionalProperties:
          $ref: '#/definitions/models.Permission'
        type: object
      project_id:
        type: string
      role:
        type: string
      updated_timestamp_utc:
        type: string
    type: object
  roles_permissions.UpdatePermissionRequest:
    properties:
      actions:
        items:
          $ref: '#/definitions/models.Action'
        type: array
      resource:
        type: string
    type: object
...
host: localhost:8002
info:
  contact: {}
paths:
...
  /projects/{project_id}/roles:
    delete:
      consumes:
      - application/json
      description: Deletes all roles for a specific project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete permissions by project
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Gets all roles for a specific project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/roles_permissions.RoleResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get roles by project
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Creates a new role
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Role details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/roles_permissions.RoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles_permissions.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - roles
  /projects/{project_id}/roles/{role_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a role by ID
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete permission
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Gets a role by ID
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles_permissions.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get role
      tags:
      - roles
  /projects/{project_id}/roles/{role_id}/permissions:
    put:
      consumes:
      - application/json
      description: Updates a specific attribute of a permission
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      - description: Attribute update details
        in: body
        name: attribute
        required: true
        schema:
          $ref: '#/definitions/roles_permissions.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update permission attribute
      tags:
      - permissions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"